Move Toggle from react-toggle to react-internal
