+++ packages/react-examples/src/react/Toggle/Toggle.Basic.Example.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Toggle } from '@fluentui/react-toggle/lib/next';
+import { Toggle } from '@fluentui/react-toggle/lib/Toggle';
 import { Stack, IStackTokens } from '@fluentui/react/lib/Stack';
 
 const stackTokens: IStackTokens = { childrenGap: 10 };
