+++ packages/react-toggle/src/components/Toggle/Toggle.scss
