+++ packages/react-toggle/src/components/Toggle/Toggle.types.ts
@@ -1,6 +1,8 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+
 import * as React from 'react';
 import { IStyle, ITheme } from '@fluentui/style-utilities';
-import { IRefObject, IStyleFunctionOrObject, IComponentAs } from '@fluentui/utilities';
+import { IRefObject, IComponentAs, IStyleFunctionOrObject } from '@fluentui/utilities';
 
 /**
  * {@docCategory Toggle}
@@ -17,7 +19,7 @@ export interface IToggleProps extends React.HTMLAttributes<HTMLElement>, React.R
   /**
    * Render the root element as another type.
    */
-  as?: IComponentAs<React.HTMLAttributes<HTMLElement>>;
+  as?: IComponentAs<React.HTMLAttributes<HTMLElement>> | React.ElementType;
 
   /**
    * Optional callback to access the IToggle interface. Use this instead of ref for accessing
@@ -177,3 +179,20 @@ export interface IToggleStyles {
    */
   text: IStyle;
 }
+
+/**
+ * {@docCategory Toggle}
+ */
+export interface IToggleSlots {}
+
+/**
+ * {@docCategory Toggle}
+ */
+export type IToggleSlotProps = {
+  [key in keyof IToggleSlots]: IToggleProps[key];
+};
+
+/**
+ * {@docCategory Toggle}
+ */
+export interface IToggleOptions {}
